Configurable request callback
For now, the actual implementation for executing requests is placed on
the module consumer. The only assumptions the client makes are that the
given request handler is a function and that it can make sense of the
few parameters we invoke it with.

This also ditches the fake test in the repo with a real one.